% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_packages}
\alias{check_packages}
\title{Check if Packages are Available}
\usage{
check_packages(packages)
}
\arguments{
\item{packages}{Character vector of package names}
}
\value{
Logical vector of same length
}
\description{
Check if Packages are Available
}
\keyword{internal}
