% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_clusters}
\alias{export_clusters}
\title{Export Clustering Results}
\usage{
export_clusters(obj, data, file)
}
\arguments{
\item{obj}{A clustering object}

\item{data}{Original data}

\item{file}{File path for output}
}
\description{
Export clustering results to CSV
}
