% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_centers}
\alias{extract_centers}
\title{Extract Cluster Centers}
\usage{
extract_centers(obj)
}
\arguments{
\item{obj}{A clustering object}
}
\value{
A tibble of cluster centers
}
\description{
Get cluster centers from various clustering objects
}
