% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{plot.tidy_autoencoder}
\alias{plot.tidy_autoencoder}
\title{Plot Autoencoder Results}
\usage{
\method{plot}{tidy_autoencoder}(x, color = NULL, size = 3, alpha = 0.7, ...)
}
\arguments{
\item{x}{A tidy_autoencoder object}

\item{color}{Optional vector for coloring points}

\item{size}{Point size. Default is 3}

\item{alpha}{Point transparency. Default is 0.7}

\item{...}{Additional arguments passed to ggplot2::geom_point()}
}
\value{
A ggplot2 object
}
\description{
Creates visualizations of autoencoder encoding
}
