% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.R
\name{plot.tidy_som}
\alias{plot.tidy_som}
\title{Plot Self-Organizing Map}
\usage{
\method{plot}{tidy_som}(x, type = "codes", property = NULL, ...)
}
\arguments{
\item{x}{A tidy_som object}

\item{type}{Type of plot: "codes", "counts", "quality", "mapping", or "property"}

\item{property}{If type is "property", which variable to plot}

\item{...}{Additional arguments (currently unused)}
}
\value{
A ggplot2 object
}
\description{
Creates various visualizations for SOM results
}
