% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_mds}
\alias{plot_mds}
\title{Plot MDS Results}
\usage{
plot_mds(
  mds_obj,
  data = NULL,
  cluster_col = NULL,
  dim1 = 1,
  dim2 = 2,
  title = "MDS Plot",
  show_stress = TRUE
)
}
\arguments{
\item{mds_obj}{A tidy_mds object or matrix of MDS coordinates}

\item{data}{Original data frame (optional, for adding cluster info)}

\item{cluster_col}{Name of cluster column if present in data}

\item{dim1}{First dimension to plot (default: 1)}

\item{dim2}{Second dimension to plot (default: 2)}

\item{title}{Plot title}

\item{show_stress}{Show stress value in subtitle? (default: TRUE)}
}
\value{
A ggplot object
}
\description{
Visualize MDS dimensions with optional cluster coloring
}
