% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_basket.R
\name{print.tidy_apriori}
\alias{print.tidy_apriori}
\title{Print Method for tidy_apriori}
\usage{
\method{print}{tidy_apriori}(x, ...)
}
\arguments{
\item{x}{A tidy_apriori object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for tidy_apriori
}
