% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{print.tidy_hclust}
\alias{print.tidy_hclust}
\title{Print Method for tidy_hclust}
\usage{
\method{print}{tidy_hclust}(x, ...)
}
\arguments{
\item{x}{A tidy_hclust object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for tidy_hclust
}
