% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomap.R
\name{print.tidy_isomap}
\alias{print.tidy_isomap}
\title{Print method for tidy_isomap}
\usage{
\method{print}{tidy_isomap}(x, ...)
}
\arguments{
\item{x}{A tidy_isomap object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for tidy_isomap
}
