% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{print.tidy_pca}
\alias{print.tidy_pca}
\title{Print Method for tidy_pca}
\usage{
\method{print}{tidy_pca}(x, ...)
}
\arguments{
\item{x}{A tidy_pca object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for tidy_pca
}
