% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som.R
\name{print.tidy_som}
\alias{print.tidy_som}
\title{Print method for tidy_som}
\usage{
\method{print}{tidy_som}(x, ...)
}
\arguments{
\item{x}{A tidy_som object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for tidy_som
}
