% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_to_range}
\alias{scale_to_range}
\title{Scale to Range}
\usage{
scale_to_range(x, new_min = 0, new_max = 1)
}
\arguments{
\item{x}{Numeric vector}

\item{new_min}{New minimum (default: 0)}

\item{new_max}{New maximum (default: 1)}
}
\value{
Scaled numeric vector
}
\description{
Scale numeric vector to a specific range
}
