% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suggest_packages}
\alias{suggest_packages}
\title{Suggest Missing Packages}
\usage{
suggest_packages(packages)
}
\arguments{
\item{packages}{Character vector of package names}
}
\description{
Suggest Missing Packages
}
\keyword{internal}
