% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{summary.tidy_autoencoder}
\alias{summary.tidy_autoencoder}
\title{Summary method for tidy_autoencoder}
\usage{
\method{summary}{tidy_autoencoder}(object, ...)
}
\arguments{
\item{object}{A tidy_autoencoder object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Summary method for tidy_autoencoder
}
