% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomap.R
\name{summary.tidy_isomap}
\alias{summary.tidy_isomap}
\title{Summary method for tidy_isomap}
\usage{
\method{summary}{tidy_isomap}(object, ...)
}
\arguments{
\item{object}{A tidy_isomap object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Summary method for tidy_isomap
}
