% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_summary}
\alias{tidy_summary}
\title{Tidy Summary}
\usage{
tidy_summary(obj)
}
\arguments{
\item{obj}{A tidy object (tidy_pca, tidy_kmeans, tidy_hclust, etc.)}
}
\value{
A tibble with summary statistics
}
\description{
Create a tidy summary of clustering or dimensionality reduction results
}
