% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_percent}
\alias{to_percent}
\title{Convert to Percentage}
\usage{
to_percent(x, digits = 1)
}
\arguments{
\item{x}{Numeric vector (proportion)}

\item{digits}{Number of decimal places (default: 1)}
}
\value{
Character vector with percentages
}
\description{
Convert to Percentage
}
